test: Phase 2 — migrate test suite to Vitest browser mode + SSR
- Replace karma/jasmine + jest with Vitest (node + browser projects),
  Playwright provider across chromium/firefox/webkit
- Consolidate the two overlapping RTL suites into one hook suite and
  port the useResolvedElement unit tests to vitest-browser-react
- Rewrite the SSR test without codegen: node renderToString assertion +
  self-contained browser hydration test
- Robust timing model: assert first-render snapshots and render-count
  deltas (async render() may flush the first RO measurement)
- Apply the React 19 ref-type edits (RefObject<T|null>) so the suite is
  green on both @types/react 18 and 19

Verified: full suite green on React 18.3.1 and 19.2.8; types clean on both.
