+++ src/index.ts
@@ -28,7 +28,7 @@ export type RoundingFunction = (n: number) => number;
 
 function useResizeObserver<T extends Element>(
   opts: {
-    ref?: RefObject<T> | T | null | undefined;
+    ref?: RefObject<T | null> | T | null | undefined;
     onResize?: ResizeHandler;
     box?: ResizeObserverBoxOptions;
     round?: RoundingFunction;
@@ -43,11 +43,14 @@ function useResizeObserver<T extends Element>(
   const round = opts.round || Math.round;
 
   // Using a single instance throughout the hook's lifetime
-  const resizeObserverRef = useRef<{
-    box?: ResizeObserverBoxOptions;
-    round?: RoundingFunction;
-    instance: ResizeObserver;
-  }>();
+  const resizeObserverRef = useRef<
+    | undefined
+    | {
+        box?: ResizeObserverBoxOptions;
+        round?: RoundingFunction;
+        instance: ResizeObserver;
+      }
+  >(undefined);
 
   const [size, setSize] = useState<{
     width?: number;
