+++ src/utils/useResolvedElement.ts
@@ -8,7 +8,7 @@ type SubscriberResponse = SubscriberCleanupFunction | void;
 // refs to such extent, but then composing hooks and components could not opt out of unnecessary renders.
 export default function useResolvedElement<T extends Element>(
   subscriber: (element: T) => SubscriberResponse,
-  refOrElement?: T | RefObject<T> | null,
+  refOrElement?: T | RefObject<T | null> | null,
 ): RefCallback<T> {
   const lastReportRef = useRef<{
     element: T | null;
