--- tests/.gitignore
@@ -1,1 +1,0 @@
-ssr.test.js
