+++ tests/setup.ts
@@ -1,0 +1,3 @@
+// Registers automatic cleanup between tests and the forked, retry-able
+// jest-dom matchers that Vitest browser mode ships with.
+import "vitest-browser-react";
