+++ tests/utils/awaitNextFrame.ts
@@ -1,12 +1,11 @@
-// @see https://stackoverflow.com/a/21825207/982092
-// @ts-ignore
-const isIE11 = !!window.MSInputMethodContext && !!document.documentMode;
-// TODO instead of hardcoded values, we should wait with a timeout for sizes to
-//      be reported whenever they're available. (rAF in a loop maybe)
-export default function awaitNextFrame() {
-  return new Promise((resolve) =>
-    // Seems like that on IE with the RO polyfill we need to slow things down a bit
-    // Also, 1000 / 60 did not seem to not be enough of a wait sometimes on modern browsers either.
-    setTimeout(resolve, 1000 / (isIE11 ? 5 : 30))
-  );
+// Waits for a ResizeObserver notification to be delivered and for React to
+// flush the resulting update. RO notifications are dispatched as part of the
+// browser's frame lifecycle (right before paint), so two animation frames
+// guarantee a notification has fired; the trailing macrotask then lets React's
+// scheduler commit the update. This is more reliable across engines than a
+// fixed timeout — Firefox in particular could miss a too-short wait under load.
+export default function awaitNextFrame(): Promise<void> {
+  return new Promise((resolve) => {
+    requestAnimationFrame(() => requestAnimationFrame(() => setTimeout(resolve, 0)));
+  });
 }
