+++ tests/utils/createController.ts
@@ -1,9 +1,10 @@
 // Creates a shared object which can be used in the test function, as well as
 // within the test component.
 // Provides some often used functions as well.
+import { expect } from "vitest";
 import awaitNextFrame from "./awaitNextFrame";
 
-const setSizePlaceholder: SetSizeFunction = async (params: SizeParams) => {};
+const setSizePlaceholder: SetSizeFunction = async (_params: SizeParams) => {};
 
 type SizeParams = {
   width?: number;
@@ -31,20 +32,24 @@ export default function createController() {
   let renderCount = 0;
   const incrementRenderCount = () => renderCount++;
   const assertRenderCount = (count: number) => expect(renderCount).toBe(count);
+  const getRenderCount = () => renderCount;
 
   let measuredWidth: number | undefined;
   let measuredHeight: number | undefined;
+  // The size reported on the very first render, captured synchronously before
+  // any ResizeObserver callback can fire. Under Vitest browser mode `render()`
+  // is async and may flush the first measurement before the test regains
+  // control, so post-render assertions can't observe the pre-measurement state
+  // — this snapshot can.
+  let initialSize: SizeParams | undefined;
   const reportMeasuredSize = (params: SizeParams) => {
-    if (
-      typeof params.width === "number" ||
-      typeof params.width === "undefined"
-    ) {
+    if (initialSize === undefined) {
+      initialSize = { width: params.width, height: params.height };
+    }
+    if (typeof params.width === "number" || typeof params.width === "undefined") {
       measuredWidth = params.width;
     }
-    if (
-      typeof params.height === "number" ||
-      typeof params.height === "undefined"
-    ) {
+    if (typeof params.height === "number" || typeof params.height === "undefined") {
       measuredHeight = params.height;
     }
   };
@@ -52,14 +57,20 @@ export default function createController() {
     expect(measuredWidth).toBe(params.width);
     expect(measuredHeight).toBe(params.height);
   };
+  const assertInitialSize = (params: SizeParams) => {
+    expect(initialSize?.width).toBe(params.width);
+    expect(initialSize?.height).toBe(params.height);
+  };
 
   const controller = {
     incrementRenderCount,
     assertRenderCount,
+    getRenderCount,
     reportMeasuredSize,
     assertMeasuredSize,
+    assertInitialSize,
     setSize: setSizePlaceholder,
-    provideSetSizeFunction: (ref: HTMLElement | null) => {}, // Placeholder to make TS happy
+    provideSetSizeFunction: (_ref: HTMLElement | null) => {}, // Placeholder to make TS happy
     triggerRender: async () => {},
   };
