--- tests/utils/delay.ts
@@ -1,2 +1,0 @@
-export default (time: number) =>
-  new Promise((resolve) => setTimeout(resolve, time));
